spectrogram: use double-precision trig for DFT coefficient table
Use cos()/sin() instead of cosf()/sinf() in rdft_init() so the
trig functions evaluate in full double precision before narrowing
to float on store. Avoids premature narrowing of the angle argument
which can reduce accuracy of DFT coefficients for large n.

+++ src/spectrogram.c
@@ -331,8 +331,8 @@ static float *rdft_init(size_t n)
   for (j = 0; j <= (int)(n / 2); ++j) {
     for (i = 0; i < (int)n; ++i) {
       double angle = 2 * M_PI * (double)j * i / n;
-      *p++ = cosf(angle);
-      *p++ = sinf(angle);
+      *p++ = cos(angle);
+      *p++ = sin(angle);
     }
   }
